move interfaces to new file, structuring
