+++ src/mappings/utils.ts
@@ -6,15 +6,7 @@ import { BigNumber } from '@ethersproject/bignumber';
 import { EraManager } from '@subql/contract-sdk';
 import testnetAddresses from '@subql/contract-sdk/publish/testnet.json';
 
-import {
-  Delegator,
-  Indexer,
-  EraValue,
-  JSONBigInt,
-  Exception,
-  Withdrawl,
-} from '../types';
-import { WithdrawlParams } from '../customTypes';
+import { Delegator, Indexer, EraValue, JSONBigInt, Exception } from '../types';
 
 export const QUERY_REGISTRY_ADDRESS = testnetAddresses.QueryRegistry.address;
 export const ERA_MANAGER_ADDRESS = testnetAddresses.EraManager.address;
@@ -75,29 +67,6 @@ export const operations: Record<string, (a: bigint, b: bigint) => bigint> = {
   sub: (a, b) => a - b,
   replace: (a, b) => b,
 };
-
-export async function upsertWithdrawlEntity({
-  id,
-  delegator,
-  indexer,
-  index,
-  amount,
-  claimed,
-  event,
-}: WithdrawlParams): Promise<void> {
-  const withdrawl = Withdrawl.create({
-    id,
-    delegator: delegator,
-    indexer: indexer,
-    index: index.toBigInt(),
-    startTime: event.blockTimestamp,
-    amount: amount.toBigInt(),
-    claimed,
-    createdBlock: event.blockNumber,
-  });
-
-  await withdrawl.save();
-}
 
 export async function upsertEraValue(
   eraManager: EraManager,
